Refactor functions in deed_service file Fix bugs for unit and integration test
